requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# LangChain & LangGraph (Updated to compatible newer versions)
langchain==0.3.13
langgraph==0.2.60
langchain-openai==0.2.14
langsmith==0.2.7
langchain-community==0.3.13
langchain-core==0.3.28
langgraph-checkpoint-postgres==3.0.2

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# pgvector
pgvector==0.2.4

# Async & Background Tasks
celery==5.3.6
redis==5.0.1
flower==2.0.1

# Document Processing
pypdf==4.0.0
python-docx==1.1.0
python-magic==0.4.27

# AWS S3
boto3==1.34.34

# Monitoring
prometheus-client==0.19.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic>=2.7.4,<3.0.0
pydantic-settings>=2.1.0

# HTTP Client
httpx==0.26.0

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Utilities
tenacity==8.2.3

# Modular RAG Components
sentence-transformers>=2.7.0  # For cross-encoder reranking (v2.7+ compatible with newer huggingface-hub)
